use json file for constant data
