+++ apps/web/app/clients/layout.tsx
@@ -1,6 +1,6 @@
 import { HeaderClient, User } from '@repo/ui'
 
-export default function Layout({
+export default function ClientsLayout({
   children,
 }: {
   children: React.ReactNode
